+++ packages/app/src/common/types.types.ts
@@ -1,3 +1,8 @@
+import type { EventObject } from '#events';
+import type { StateExtended } from '#states';
+import type { Decompose } from '@bemedev/decompose';
+import type { PrimitiveObject } from '@bemedev/typings';
+
 export type TimeActionsTypes =
   | 'pauseActivity'
   | 'resumeActivity'
@@ -18,3 +23,16 @@ type _ActionTypes =
 export type ActionTypes = `actions.${_ActionTypes}`;
 
 export type AppTypes = ActionTypes | 'guards' | 'pContext' | 'context';
+
+export type ByKey_F<
+  E extends EventObject = EventObject,
+  Pc = any,
+  Tc extends PrimitiveObject = PrimitiveObject,
+  T extends string = string,
+> = <
+  S extends StateExtended<E, Pc, Tc, T>,
+  D = Decompose<S, { object: 'both'; start: false; sep: '.' }>,
+  K extends keyof D & string = keyof D & string,
+>(
+  key: K,
+) => D[K];